Refactor out point cloud def
