+++ src/perception/point_cloud.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <cstdint>
+
+namespace mrover {
+
+    struct Point {
+        float x, y, z;
+        uint8_t r, g, b, a;
+        float normal_x, normal_y, normal_z;
+        float curvature;
+    } __attribute__((packed));
+
+} // namespace mrover
